edit hw

+++ src/controllers/contacts.js
@@ -55,18 +55,22 @@ export const getContactByIdController = async (req, res, next) => {
 export const createContactController = async (req, res, next) => {
   let photo = null;
 
-  if (getEnvVar('UPLOAD_TO_CLOUDINARY') === 'true') {
-    const result = await saveFileToCloudinary(req.file.path);
-    await fs.unlink(req.file.path);
+  if (req.file) {
+    if (getEnvVar('UPLOAD_TO_CLOUDINARY') === 'true') {
+      const result = await saveFileToCloudinary(req.file.path);
+      await fs.unlink(req.file.path);
 
-    photo = result.secure_url;
+      photo = result.secure_url;
+    } else {
+      await fs.rename(
+        req.file.path,
+        path.resolve('src', 'uploads', 'photos', req.file.filename),
+      );
+
+      photo = `http://localhost:3000/photo/${req.file.filename}`;
+    }
   } else {
-    await fs.rename(
-      req.file.path,
-      path.resolve('src', 'uploads', 'photos', req.file.filename),
-    );
-
-    photo = `http://localhost:3000/photo/${req.file.filename}`;
+    photo = null;
   }
 
   const contact = await createContact({
